Create new Image entity and deleted Artwork y ArtworkKind
