+++ Jellyfin.Data/Entities/AccessSchedule.cs
@@ -1,4 +1,4 @@
-﻿using System;
+using System;
 using System.ComponentModel.DataAnnotations.Schema;
 using System.Xml.Serialization;
 using Jellyfin.Data.Enums;
